backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
aiomysql==0.2.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.5.2
pydantic[email]==2.9.2

# File uploads
python-multipart==0.0.12

# Redis (token blacklist & idempotency in production)
redis==5.1.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
